Intermediary commit of testing code and some refactoring
Intermediary commit of testing code and some refactoring
