+++ python/parse_sql/config.yaml
@@ -3,8 +3,8 @@ SQL_SERVER: "localhost,2017"
 SQL_DATABASE: "parse_sql"
 SQL_LOGIN: "python_user"
 SQL_PASSWORD: "python_user"
-# Directory where the .sql files are
-SQL_DIR: "C:\\Users\\jong\\Documents\\GitHub\\Rave\\Medidata 5 RAVE Database Project"
+# Directory where to recursively look for .sql files
+SQL_DIR: "C:\\Users\\jong\\Documents\\GitHub\\Rave"
 # Grab the current git git_tag from repo
 GIT_REPO_MAIN_DIR: "C:\\Users\\jong\\Documents\\GitHub\\Rave"
 GIT_REPO_NAME: "Rave"
