+++ python/parse_sql/test/__init__.py
@@ -1,0 +1,1 @@
+"""Initialization of package."""
