+++ python/parse_sql/test/ddl_examples/ddl_function_assertions.json
@@ -1,0 +1,1 @@
+["CREATE FUNCTION dbo.ISOweek", "CREATE FUNCTION Sales.ufn_SalesByStore", "CREATE FUNCTION dbo.ufn_FindReports", "CREATE FUNCTION dbo.ufnGetInventoryStock", "CREATE FUNCTION Sales.ufn_SalesByStore", "CREATE FUNCTION dbo.ufn_FindReports", "DROP FUNCTION ufnGetInventoryStock", "DROP FUNCTION Sales.ufn_SalesByStore", "DROP FUNCTION dbo.ufn_FindReports", "ALTER FUNCTION [dbo].[ufnGetContactInformation]", "ALTER FUNCTION Sundays"]
