+++ python/parse_sql/test/ddl_examples/ddl_index_assertions.json
@@ -1,0 +1,1 @@
+["CREATE INDEX i1 ON t1 (col1)", "CREATE CLUSTERED INDEX index1 ON d1.s1.t1 (col1)", "CREATE UNIQUE INDEX unique1 ON t1 (col1 DESC, col2 ASC, col3 DESC)", "CREATE UNIQUE CLUSTERED INDEX Idx1 ON t1(c)", "CREATE INDEX IX_VendorID ON ProductVendor (VendorID)", "CREATE INDEX IX_VendorID ON dbo.ProductVendor (VendorID DESC, Name ASC, Address DESC)", "CREATE INDEX IX_VendorID ON Purchasing..ProductVendor (VendorID)", "CREATE NONCLUSTERED INDEX IX_SalesPerson_SalesQuota_SalesYTD ON Sales.SalesPerson (SalesQuota, SalesYTD)", "CREATE INDEX IX_FF ON dbo.FactFinance ( FinanceKey ASC, DateKey ASC )", "CREATE INDEX IX_FF ON dbo.FactFinance ( FinanceKey, DateKey, OrganizationKey DESC)", "CREATE NONCLUSTERED INDEX IX_ProductVendor_VendorID   \n    ON Purchasing.ProductVendor (BusinessEntityID)", "CREATE UNIQUE CLUSTERED INDEX  \n    AK_BillOfMaterials_ProductAssemblyID_ComponentID_StartDate   \n        ON Production.BillOfMaterials (ProductAssemblyID, ComponentID,   \n        StartDate)", "CREATE CLUSTERED INDEX cci_SimpleTable ON SimpleTable (ProductKey)", "DROP INDEX IX_ProductVendor_VendorID ON Purchasing.ProductVendor", "DROP INDEX IX_ProductVendor_BusinessEntityID   \n    ON Purchasing.ProductVendor", "DROP INDEX PXML_ProductModel_CatalogDescription   \n    ON Production.ProductModel", "DROP INDEX  \n    IX_PurchaseOrderHeader_EmployeeID ON Purchasing.PurchaseOrderHeader", "DROP INDEX AK_BillOfMaterials_ProductAssemblyID_ComponentID_StartDate   \n    ON Production.BillOfMaterials"]
