+++ python/parse_sql/test/ddl_examples/ddl_procedure_assertions.json
@@ -1,0 +1,1 @@
+["CREATE PROCEDURE dbo.usp_add_kitchen", "create    procedure    SelectAllCustomers", "CREATE PROCEDURE Purchasing.uspVendorAllInfo", "DROP PROCEDURE dbo.uspMyProc", "DROP PROCEDURE IF EXISTS dbo.uspMyProc3", "DROP PROCEDURE dbo.uspGetSalesbyMonth", "DROP PROCEDURE MyProc", "DROP PROCEDURE Purchasing.uspVendorAllInfo", "CREATE PROC What_DB_is_this", "CREATE PROC[dbo].[What_DB_is_this2]", "Create PROC What_DB_is_that", "DROP PROC dbo.uspMyProc2", "ALTER PROCEDURE Purchasing.uspVendorAllInfo", "ALTER PROC Purchasing.uspVendorAllInfo"]
