+++ python/parse_sql/test/ddl_examples/ddl_rename_assertions.json
@@ -1,0 +1,1 @@
+["sp_rename 'Sales.SalesTerritory', 'SalesTerr'", "sp_rename 'Sales.SalesTerritory.TerritoryID', 'TerrID'", "sp_rename N'Purchasing.ProductVendor.IX_ProductVendor_VendorID', N'IX_VendorID'", "sp_rename N'Phone', N'Telephone'", "sp_rename 'HumanResources.PK_Employee_BusinessEntityID', 'PK_EmployeeID'", "sp_rename 'HumanResources.CK_Employee_BirthDate', 'CK_BirthDate'", "sp_rename 'HumanResources.FK_Employee_Person_BusinessEntityID', 'FK_EmployeeID'", "sp_rename @objname = 'tableA' ,  @newname = 'tableB'   ", "sp_rename @objname = 'tableC' ,  @newname = 'tableD'   "]
