+++ python/parse_sql/test/ddl_examples/ddl_rename_examples.sql
@@ -1,0 +1,58 @@
+-- =============================================
+-- SP_RENAME
+-- =============================================
+
+USE AdventureWorks2012;  
+GO  
+EXEC sp_rename 'Sales.SalesTerritory', 'SalesTerr';  
+GO  
+
+USE AdventureWorks2012;  
+GO  
+EXEC sp_rename 'Sales.SalesTerritory.TerritoryID', 'TerrID', 'COLUMN';  
+GO  
+
+USE AdventureWorks2012;  
+GO  
+EXEC sp_rename N'Purchasing.ProductVendor.IX_ProductVendor_VendorID', N'IX_VendorID', N'INDEX';  
+GO  
+
+USE AdventureWorks2012;  
+GO  
+EXEC sp_rename N'Phone', N'Telephone', N'USERDATATYPE';  
+GO  
+
+-- Rename the primary key constraint.  
+sp_rename 'HumanResources.PK_Employee_BusinessEntityID', 'PK_EmployeeID';  
+GO  
+
+-- Rename a check constraint.  
+sp_rename 'HumanResources.CK_Employee_BirthDate', 'CK_BirthDate';  
+GO  
+
+-- Rename a foreign key constraint.  
+sp_rename 'HumanResources.FK_Employee_Person_BusinessEntityID', 'FK_EmployeeID';  
+
+sp_rename @objname = 'tableA' ,  @newname = 'tableB'   
+
+sp_rename @objname = 'tableC' ,  @newname = 'tableD'   
+     , @objtype = 'TABLE'  
+
+
+
+
+
+
+
+
+
+
+--CONSTRAINTS
+
+
+-- Set ONLINE = OFF to execute this example on editions other than Enterprise Edition.  
+ALTER TABLE Production.TransactionHistoryArchive  
+DROP CONSTRAINT PK_TransactionHistoryArchive_TransactionID  
+WITH (ONLINE = ON);  
+
+
