+++ python/parse_sql/test/ddl_examples/ddl_table_assertions.json
@@ -1,0 +1,1 @@
+["CREATE TABLE Persons", "CREATE TABLE dbo.SUPPLY1", "create table dbo.SUPPLY2", "CREATE\t TABLE   \n dbo.SUPPLY3", "CREATE TABLE \n\ndbo.SUPPLY4", "CREATE TABLE t1", "DROP TABLE Shippers", "DROP TABLE [dbo].[Shippers2]", "INTO dbo.CustomersBackup2019\nFROM", "INTO CustomersOrderBackup2020\nFROM", "INTO CustomersBackup2017\nFROM", "INTO Fast_Customers from", "INTO dbo.DepartmentsUsingOpenDataSource  \nFROM", "INTO dbo.Departments  \nFROM", "INTO dbo.DepartmentsUsingOpenQuery  \nFROM", "INTO Person.USAddress   \nFROM", "INTO dbo.NewProducts  \nFROM", "INTO dbo.EmployeeAddresses  \nFROM", "ALTER TABLE Persons\nADD DateOfBirth date", "ALTER \nTABLE\ndbo.[t1] \nADD \ncol3 \nvarchar\n(100)\n\n", "ALTER TABLE Persons\nDROP COLUMN DateOfBirth3", "ALTER TABLE Customer DROP COLUMN Birth_Date", "ALTER TABLE dbo.Customer \nDROP COLUMN Birth_Date2", "ALTER TABLE Persons\nALTER COLUMN DateOfBirth2 year", "ALTER TABLE dbo.[tbl] ALTER COLUMN col1 varchar(300) NULL", "ALTER TABLE dbo.[tbl] ALTER COLUMN col2 varchar(20) NOT NULL", "ALTER \nTABLE \ndbo.[SUPPLY4] \nALTER \nCOLUMN \nsupplier \nvarchar(50)\nNOT NULL\n\n", "ALTER TABLE dbo.[tbl] ALTER COLUMN col4 varchar(100)\n\n"]
