+++ python/parse_sql/test/ddl_examples/ddl_trigger_examples.sql
@@ -1,0 +1,117 @@
+-- =============================================
+-- TRIGGERS
+-- =============================================
+
+CREATE TRIGGER Purchasing.LowCredit ON Purchasing.PurchaseOrderHeader  
+AFTER INSERT  
+AS  
+IF (@@ROWCOUNT_BIG  = 0)
+RETURN;
+IF EXISTS (SELECT *  
+           FROM Purchasing.PurchaseOrderHeader AS p   
+           JOIN inserted AS i   
+           ON p.PurchaseOrderID = i.PurchaseOrderID   
+           JOIN Purchasing.Vendor AS v   
+           ON v.BusinessEntityID = p.VendorID  
+           WHERE v.CreditRating = 5  
+          )  
+BEGIN  
+RAISERROR ('A vendor''s credit rating is too low to accept new  
+purchase orders.', 16, 1);  
+ROLLBACK TRANSACTION;  
+RETURN   
+END;  
+GO  
+
+CREATE TRIGGER safety   
+ON DATABASE   
+FOR DROP_SYNONYM  
+AS   
+IF (@@ROWCOUNT = 0)
+RETURN;
+   RAISERROR ('You must disable Trigger "safety" to drop synonyms!',10, 1)  
+   ROLLBACK  
+GO  
+DROP TRIGGER safety  
+ON DATABASE;  
+GO  
+
+
+CREATE TRIGGER Sales.bonus_reminder  
+ON Sales.SalesPersonQuotaHistory  
+WITH ENCRYPTION  
+AFTER INSERT, UPDATE   
+AS RAISERROR ('Notify Compensation', 16, 10);  
+GO  
+
+-- Now, change the trigger.  
+ALTER TRIGGER Sales.bonus_reminder  
+ON Sales.SalesPersonQuotaHistory  
+AFTER INSERT  
+AS RAISERROR ('Notify Compensation', 16, 10);  
+GO  
+
+ALTER TRIGGER trgAfterInsert ON [dbo].[Employee_Test] 
+FOR INSERT
+AS
+	declare @empid int;
+	declare @empname varchar(100);
+	declare @empsal decimal(10,2);
+	declare @audit_action varchar(100);
+
+	select @empid=i.Emp_ID from inserted i;	
+	select @empname=i.Emp_Name from inserted i;	
+	select @empsal=i.Emp_Sal from inserted i;	
+	set @audit_action='Inserted Record -- After Insert Trigger.';
+
+	insert into Employee_Test_Audit
+           (Emp_ID,Emp_Name,Emp_Sal,Audit_Action,Audit_Timestamp) 
+	values(@empid,@empname,@empsal,@audit_action,getdate());
+
+	PRINT 'AFTER INSERT trigger fired.'
+GO
+
+ALTER TRIGGER trgAfterUpdate ON [dbo].[Employee_Test] 
+FOR UPDATE
+AS
+	declare @empid int;
+	declare @empname varchar(100);
+	declare @empsal decimal(10,2);
+	declare @audit_action varchar(100);
+
+	select @empid=i.Emp_ID from inserted i;	
+	select @empname=i.Emp_Name from inserted i;	
+	select @empsal=i.Emp_Sal from inserted i;	
+	
+	if update(Emp_Name)
+		set @audit_action='Updated Record -- After Update Trigger.';
+	if update(Emp_Sal)
+		set @audit_action='Updated Record -- After Update Trigger.';
+
+	insert into Employee_Test_Audit(Emp_ID,Emp_Name,Emp_Sal,Audit_Action,Audit_Timestamp) 
+	values(@empid,@empname,@empsal,@audit_action,getdate());
+
+	PRINT 'AFTER UPDATE Trigger fired.'
+GO
+
+ALTER TRIGGER MyTrigger
+ON MySchema.MyTable
+AFTER INSERT, UPDATE
+AS
+   EXEC msdb.dbo.sp_send_dbmail
+        @profile_name = 'Database Administrator',
+        @recipients = 'name@somewhere.com',
+        @body = 'email message goes here.',
+        @subject = 'New record added';
+GO
+
+IF OBJECT_ID ('TRG_PREVENT_VIEW_DROP') IS NOT NULL
+DROP TRIGGER TRG_PREVENT_VIEW_DROP
+
+GO
+
+IF OBJECT_ID ('TRG_PREVENT_VIEW_DROP') IS NOT NULL
+DROP 
+TRIGGER [dbo].TRG_PREVENT_VIEW_DROP
+
+GO
