+++ python/parse_sql/test/ddl_examples/ddl_view_examples.sql
@@ -1,0 +1,96 @@
+-- =============================================
+-- VIEWS
+-- =============================================
+
+CREATE VIEW [Products Above Average Price] AS
+SELECT ProductName, UnitPrice
+FROM Products
+WHERE UnitPrice > (SELECT AVG(UnitPrice) FROM Products);
+
+CREATE VIEW hiredate_view  
+AS   
+SELECT p.FirstName, p.LastName, e.BusinessEntityID, e.HireDate  
+FROM HumanResources.Employee e   
+JOIN Person.Person AS p ON e.BusinessEntityID = p.BusinessEntityID ;  
+GO  
+
+CREATE VIEW Purchasing.PurchaseOrderReject  
+WITH ENCRYPTION  
+AS  
+SELECT PurchaseOrderID, ReceivedQty, RejectedQty,   
+    RejectedQty / ReceivedQty AS RejectRatio, DueDate  
+FROM Purchasing.PurchaseOrderDetail  
+WHERE RejectedQty / ReceivedQty > 0  
+AND DueDate > CONVERT(DATETIME,'20010630',101) ;  
+GO  
+
+CREATE VIEW dbo.SeattleOnly  
+AS  
+SELECT p.LastName, p.FirstName, e.JobTitle, a.City, sp.StateProvinceCode  
+FROM HumanResources.Employee e  
+INNER JOIN Person.Person p  
+ON p.BusinessEntityID = e.BusinessEntityID  
+    INNER JOIN Person.BusinessEntityAddress bea   
+    ON bea.BusinessEntityID = e.BusinessEntityID   
+    INNER JOIN Person.Address a   
+    ON a.AddressID = bea.AddressID  
+    INNER JOIN Person.StateProvince sp   
+    ON sp.StateProvinceID = a.StateProvinceID  
+WHERE a.City = 'Seattle'  
+WITH CHECK OPTION ;  
+GO  
+
+
+CREATE VIEW Sales.SalesPersonPerform  
+AS  
+SELECT TOP (100) SalesPersonID, SUM(TotalDue) AS TotalSales  
+FROM Sales.SalesOrderHeader  
+WHERE OrderDate > CONVERT(DATETIME,'20001231',101)  
+GROUP BY SalesPersonID;  
+GO  
+
+--Partitioned view as defined on Server1  
+CREATE VIEW Customers  
+AS  
+--Select from local member table.  
+SELECT *  
+FROM CompanyData.dbo.Customers_33  
+UNION ALL  
+--Select from member table on Server2.  
+SELECT *  
+FROM Server2.CompanyData.dbo.Customers_66  
+UNION ALL  
+--Select from mmeber table on Server3.  
+SELECT *  
+FROM Server3.CompanyData.dbo.Customers_99; 
+
+
+ALTER VIEW HumanResources.EmployeeHireDate  
+AS  
+SELECT p.FirstName, p.LastName, e.HireDate  
+FROM HumanResources.Employee AS e JOIN Person.Person AS p  
+ON e.BusinessEntityID = p.BusinessEntityID  
+WHERE HireDate < CONVERT(DATETIME,'20020101',101) ;  
+GO  
+
+	alter VIEW 
+	emp_view
+AS 
+SELECT c.FirstName, c.LastName, s.StaffID
+FROM MySchema.Staff S 
+JOIN MySchema.Customers c 
+on c.CustomerID = c.StaffID ;
+GO
+
+ALTER  VIEW [dbo].[tbl]
+AS
+SELECT col1, col2, col3, col4
+FROM   dbo.[tbl]
+GO
+
+DROP VIEW view_name;
+
+drop view dbo.asdf;
+
+ drop View [dbo].[my_view_name];
+
