+++ python/read_csv_to_mssql/read_csv_to_mssql.py
@@ -61,11 +61,10 @@ db_indexes.to_sql("database_indexes",
                          'table_name': sqlalchemy.types.NVARCHAR(length=255),
                          'index_id': sqlalchemy.types.INT(),
                          'type_desc': sqlalchemy.types.NVARCHAR(length=255),
-                         'is_unique': sqlalchemy.types.BIT(),
-                         'is_primary_key': sqlalchemy.types.BIT(),
-                         'is_unique_constraint': sqlalchemy.types.BIT(),
+                         'is_unique': sqlalchemy.types.SMALLINT(),
+                         'is_primary_key': sqlalchemy.types.SMALLINT(),
+                         'is_unique_constraint': sqlalchemy.types.SMALLINT(),
                          'url': sqlalchemy.types.NVARCHAR(length=500),
                          'urltype': sqlalchemy.types.NVARCHAR(length=50),
                          'sqlserver': sqlalchemy.types.NVARCHAR(length=255),
                          'dbname': sqlalchemy.types.NVARCHAR(length=255)})
-s
